+++ src/components/modal/modal.module.d.ts
@@ -1,0 +1,7 @@
+declare module "*/modal.module.css" {
+const styles: {
+"modal": string
+"content": string
+}
+export default styles
+}